moved readImageInfo
