+++ Modules/Filtering/LabelErodeDilate/test/itkLabelSetErodeTest.cxx
@@ -4,6 +4,9 @@
 
 #include "itkLabelSetErodeImageFilter.h"
 
+int readImageInfo(std::string filename, itk::ImageIOBase::IOComponentType *ComponentType, int *dim);
+
+#include "read_info.h"
 
 template <class MaskPixType, int dim>
 int doErode(char *In, char *Out, int radius)
@@ -48,21 +51,7 @@ int doErode(char *In, char *Out, int radius)
 
 }
 
-/////////////////////////////////
-int readImageInfo(std::string filename, itk::ImageIOBase::IOComponentType *ComponentType, int *dim)
-{
-  itk::ImageIOBase::Pointer imageIO = itk::ImageIOFactory::CreateImageIO(filename.c_str(), itk::ImageIOFactory::ReadMode);
-  if (imageIO.IsNull())
-    return 0;
-
-
-  imageIO->SetFileName(filename.c_str());
-  imageIO->ReadImageInformation();
-
-  *ComponentType = imageIO->GetComponentType();
-  *dim = imageIO->GetNumberOfDimensions();
-  return(1);
-}
+/////////////////////////////////////////////
 
 int itkLabelSetErodeTest(int argc, char * argv[])
 {
